+++ Mc2.CrudTest.Domain/Dtos/Invoices/InvoiceInputDto.cs
@@ -6,7 +6,7 @@ namespace Mc2.CrudTest.Domain.Dtos.Invoices
     public class InvoiceInputDto
     {
         public Guid? InvoiceId { get; set; }
-        public int InvoiceNumber { get; set; }
+        public long InvoiceNumber { get; set; }
         public string ExternalInvoiceNumber { get; set; }
         public Status InvoiceStatus { get; set; }
         public decimal TotalAmount { get; set; }
